Remove spurious space token and simplify \renew@leipzig
The "\if\relax\detokenize{#2}\relax" test was incorrect because #1 was
(most probably) meant instead of #2. No one was hit by this error
however, because this test appears to be unneeded. Indeed, passing an
empty argument to \newleipzig doesn't seem to cause any problem: the
optional argument of glossaries.sty's \newacronym simply ends with a
comma in this case, which works fine, as far as I can see.