Remove an unwanted \@gobbletwo
This \@gobbletwo in the replacement text of \renewleipzig was apparently
copied from the case where glossaries.sty is loaded, however things are
different here. Indeed, when glossaries.sty is loaded, \renewleipzig
grabs two arguments (\newcommand{\renewleipzig}[2][]{...}), which is why
two more "arguments" must be gobbled in case the abbreviation is
undefined; however, when glossaries.sty is *not* loaded, \renewleipzig
grabs *four* arguments (\newcommand{\renewleipzig}[4][]{...}), therefore
if it has to issue the \PackageError, no more arguments need to be
gobbled.